refactor: simplify source structure
